third_party/blink/web_tests/platform/linux-chrome/wpt_internal/hid/requestDevice/sandboxed_iframe.https.window-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: assert_equals: expected (string) "Success" but got (object) object "[object Object]"
[TIMEOUT] Calls to HID APIs from a sandboxed iframe are valid.
  Test timed out
Harness: the test ran to completion.
